% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/myvif.R
\name{myvif}
\alias{myvif}
\title{Support function for corvif. Will not be called by the user
\code{myvif}}
\usage{
myvif(mod)
}
\description{
Support function for corvif. Will not be called by the user
\code{myvif}
}

