% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel-smooth-2.R
\name{panel.smooth2}
\alias{panel.smooth2}
\title{panel.smooth2}
\usage{
panel.smooth2(x, y, col = par("col"), bg = NA, pch = par("pch"),
  cex = 1, col.smooth = "black", span = 2/3, iter = 3, ...)
}
\description{
panel.smooth2
}

